requirements.txt
torch
transformers
whisper
gradio
gtts
Pillow
numpy
